+++ gitlab/etc/links.ignore
@@ -1,4 +1,3 @@
-# links starting with the following patterns will be ignored by linkchecker
 linkedin.com/share
 reddit.com/submit
 twitter.com/share
@@ -9,5 +8,6 @@ file:
 localhost:
 url.to.snpashot.com
 app.datadoghq.com
+app.datadog.com
 some-account.atlassian.net
 bitbucket.org/groupname/reponame
